LiquidityProviderAgent/src/LiquidityProviderAgent/agent.py

from autogen import AssistantAgent
from agoric_sdk import ContractAPI

class LiquidityProviderAgent:
    def __init__(self):
        self.assistant = AssistantAgent("liquidity_assistant")
        self.contract_api = ContractAPI("https://api.agoric.net")

    def monitor_pools(self):
        try:
            # Monitor the liquidity pools
            pools = self.contract_api.get_liquidity_pools()
            return pools
        except Exception as e:
            return f"Failed to monitor pools: {str(e)}"

    def provide_liquidity(self, pool_id, amount):
        try:
            # Provide liquidity to the specified pool
            result = self.contract_api.add_liquidity(pool_id, amount)
            return result
        except Exception as e:
            return f"Failed to provide liquidity: {str(e)}"

    def withdraw_liquidity(self, pool_id, amount):
        try:
            # Withdraw liquidity from the specified pool
            result = self.contract_api.remove_liquidity(pool_id, amount)
            return result
        except Exception as e:
            return f"Failed to withdraw liquidity: {str(e)}"

if __name__ == "__main__":
    agent = LiquidityProviderAgent()
    pools = agent.monitor_pools()
    print(pools)
    pool_id = "example_pool_id"
    amount = 100
    add_result = agent.provide_liquidity(pool_id, amount)
    print(add_result)
    remove_result = agent.withdraw_liquidity(pool_id, amount)
    print(remove_result)
